backend/requirements.txt
absl-py==2.4.0
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==26.1.0
audioread==3.1.0
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
cachetools==7.0.5
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
decorator==5.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
flatbuffers==25.12.19
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kapre==0.4.0
keras==3.13.2
kombu==5.6.2
lazy-loader==0.5
libclang==18.1.1
librosa==0.11.0
llvmlite==0.46.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
motor==3.7.1
moviepy==2.2.1
msgpack==1.1.2
namex==0.1.0
narwhals==2.18.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
pooch==1.9.0
proglog==0.1.12
prompt_toolkit==3.0.52
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pymongo==4.16.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.5
resampy==0.2.2
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
soundfile==0.13.1
soxr==1.0.0
starlette==0.52.1
streamlit==1.55.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.3.3
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.6
wrapt==2.1.2
